requirements.txt
albucore==0.0.24
albumentations==2.0.7
annotated-types==0.7.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
fsspec==2025.3.2
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
packaging==23.0
pexpect==4.8.0
pillow==11.0.0
pycocotools==2.0.8
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.4
sympy==1.13.1
torch==2.5.1
torchmetrics==1.6.3
torchvision==0.20.1
triton==3.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
